Changed executable file to LINUX format, so now everything will work only on linux

+++ server/submissionsFolder/code24.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(){
-  int a, b; cin >> a >> b;
+  long long a, b; cin >> a >> b;
   cout << a + b;
   return 0;
 }
